basi2022-main/requirements.txt
click==8.1.3
colorama==0.4.4
dominate==2.6.0
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.1
Flask-SQLAlchemy==2.5.1
Flask==2.1.2
greenlet==1.1.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pip==21.3.1
setuptools==60.2.0
SQLAlchemy==1.4.36
visitor==0.1.3
Werkzeug==2.1.2
wheel==0.37.1
python-dotenv
psycopg2-binary==2.9.3

